requirements.txt
# Root requirements for local development only
# Docker containers use docker/requirements/*.txt

tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
ipykernel==6.25.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1
Pillow==10.0.0
opencv-python==4.8.0.74
flask==2.3.3
pyswarm==0.6
requests==2.31.0
